refactor(Configs): :wrench: Refactor code and implement configurable options for increased flexibility
This commit introduces a significant update to the code and configuration files. The changes aim to make the code more versatile and easily customizable for users who want to incorporate it into their own projects.  To achieve this, the code has been refactored to implement configurable options that can be easily adjusted by users. Specifically, the following values have been modified in the configuration file:  "mastodon": an object that now takes the "client_key", "client_secret", and "access_token" values as inputs for Mastodon API authentication. "aws": an object that now takes "key", "secret", "region", and "bucket" values as inputs for AWS S3 authentication. "reddit": another object that now takes "sub" for the name of the subreddit (without the "r/") and "limit" for the number of posts to fetch on said subreddit. "saved_after": a bool that now saves or does not save the post as posted in the database. "DBUri": a string that now specifies the login information to access the mongo database.

BREAKING: Default hardcoded values.